[REFACTOR] #287 패키지 및 메소드 명 수정
